requirements.txt
requests
browser-use
scrapybara
playwright
selenium
transformers
torch
openai
pandas
numpy
faker
python-dotenv
groq